Move validateBase32 to cryptography
