+++ elements/lisk-cryptography/src/keys.ts
@@ -107,6 +107,11 @@ const polymod = (uint5Array: number[]): number => {
 	return chk;
 };
 
+export const getBinaryAddressFromPublicKey = (publicKey: string): Buffer => {
+	const publicKeyBuffer = Buffer.from(publicKey, 'hex');
+	return hash(publicKeyBuffer).slice(0, 20);
+};
+
 export const createChecksum = (uint5Array: number[]): number[] => {
 	const values = uint5Array.concat([0, 0, 0, 0, 0, 0]);
 	// eslint-disable-next-line no-bitwise
@@ -122,22 +127,51 @@ export const createChecksum = (uint5Array: number[]): number[] => {
 export const verifyChecksum = (integerSequence: number[]): boolean =>
 	polymod(integerSequence) === 1;
 
-export const getBinaryAddressFromPublicKey = (publicKey: string): Buffer => {
-	const publicKeyBuffer = Buffer.from(publicKey, 'hex');
-	return hash(publicKeyBuffer).slice(0, 40);
-};
-
 export const getBase32AddressFromPublicKey = (
 	publicKey: string,
 	prefix: string,
 ): string => {
 	const binaryAddress = getBinaryAddressFromPublicKey(publicKey);
 	const byteSequence = [];
-	for (const b of Buffer.from(binaryAddress.toString('hex'), 'hex')) {
+	for (const b of binaryAddress) {
 		byteSequence.push(b);
 	}
 	const uint5Address = convertUIntArray(byteSequence, 8, 5);
 	const uint5Checksum = createChecksum(uint5Address);
 
 	return `${prefix}${convertUInt5ToBase32(uint5Address.concat(uint5Checksum))}`;
+};
+
+const BASE32_ADDRESS_LENGTH = 41;
+const BASE32_CHARSET = 'zxvcpmbn3465o978uyrtkqew2adsjhfg';
+
+export const validateBase32Address = (address: string): boolean => {
+	if (address.length !== BASE32_ADDRESS_LENGTH) {
+		throw new Error(
+			'Address length does not match requirements. Expected 41 characters.',
+		);
+	}
+
+	const prefix = address.substring(0, 3);
+
+	if (prefix !== 'lsk') {
+		throw new Error('Invalid prefix. Expected prefix: `lsk`');
+	}
+	const addressSubstringArray = address.substring(3).split('');
+
+	if (!addressSubstringArray.every(char => BASE32_CHARSET.includes(char))) {
+		throw new Error(
+			`Invalid character found in address. Only allow characters: 'abcdefghjkmnopqrstuvwxyz23456789'.`,
+		);
+	}
+
+	const integerSequence = addressSubstringArray.map(char =>
+		BASE32_CHARSET.indexOf(char),
+	);
+
+	if (!verifyChecksum(integerSequence)) {
+		throw new Error(`Invalid checksum for address.`);
+	}
+
+	return true;
 };
